Moving the Key Command method to appropriate class
