+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController+BraveTalk.swift
@@ -33,12 +33,6 @@ extension BrowserViewController {
 #endif
   }
   
-  var braveTalkKeyCommands: [UIKeyCommand] {
-    [
-      .init(input: "m", modifierFlags: [], action: #selector(BrowserViewController.toggleBraveTalkMute))
-    ]
-  }
-  
   @objc private func toggleBraveTalkMute() {
 #if canImport(BraveTalk)
     braveTalkJitsiCoordinator.toggleMute()
